new JsonParsing class and Definitions.h

+++ Classes/Units/Character.h
@@ -27,7 +27,7 @@ protected:
 	float stamina;
 	int damage;
 	int damageRange;
-	int velocityMax;
+	float velocityMax;
 };
 
 #endif
